Downloads/Desenvolvimento Front-End/ProjetoSeparadoIA/requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
obs-websocket-py>=0.6.3
groq>=0.0.0   # opcional: groq SDK; se der problema, usamos requests
